+++ changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Version 0.2.0 (Sep 9, 2022)
+
+- Added `pack` command
+- Added yaml configuration
+- Added data packing
+- Edited the script example, added variables to it
+- Added dynamic example yaml config
+
+---
+
 ## Version 0.1.0 (Sep 9, 2022)
 
 - Added `add` command